ACN: Pseudo Optimization on simulate Annealing.

+++ Core/src/main/java/com/polytech4A/cuttingstock/core/method/Result.java
@@ -57,7 +57,7 @@ public class Result {
         this.cost = cost;
     }
 
-    public double getCost() {
+    public int getCost() {
         return cost;
     }
